disable favorite for guest

+++ app/Http/Controllers/FavoriteController.php
@@ -9,28 +9,38 @@ use Illuminate\Support\Facades\Auth;
 
 class FavoriteController extends Controller
 {
-    public function add_favorites(Request $request, $id){
-        $user_id = $request->user()->id;
-        $favorite = new Favorite();
-        $favorite->user_id = $user_id;
-        $favorite->property_id = $id;
-        $favorite->save();
-        return response()->json(['success' => 'Product added to favorites']);
+
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum');
     }
 
-    public function remove_favorites(Request $request, $id){
-        $user_id = $request->user()->id;
+    public function toggle_favorites($id){
+        $user_id = Auth::id();
+
         $favorite = Favorite::where('user_id', $user_id)->where('property_id', $id)->first();
-        $favorite->delete();
-        return response()->json(['success' => 'Product removed from favorites']);
+
+        if($favorite){
+            $favorite->delete();
+        }else{
+            $favorite = new Favorite();
+            $favorite->user_id = $user_id;
+            $favorite->property_id = $id;
+            $favorite->save();
+        }
+        return response()->json(['success' => 'Property update.']);
     }
 
-    public function show_favorites(Request $request){
-        $user_id = $request->user()->id;
+    public function show_favorites(){
+        $user_id = Auth::id();
         $properties = Property::whereHas('favorites', function($q) use ($user_id){
             $q->where('user_id', $user_id);
         })->with('images')->get();
-        return response()->json(['properties' => $properties]);
+
+        foreach ($properties as $property) {
+            $property->is_favorite = true;
+        }
+        return $properties;
     }
 
 }
